Moving and deleting files.
