+++ src/App.vue
@@ -15,29 +15,20 @@
     </v-app>
 </template>
 <script>
-
-
     import Notifications from '@/components/Notifications';
     import Navbar from './components/Navbar';
-    import LoadingApp from '@/views/LoadingApp';
     import RememberWalletModal from "@/components/modals/RememberWallet/RememberWalletModal";
 
     export default {
         components: {
             RememberWalletModal,
-
             Navbar,
             Notifications,
-            LoadingApp
         },
         async created() {
-            let parent = this;
             await this.$store.dispatch('Network/init');
     },
     computed: {
-        appReady() {
-            return this.$store.getters['appReady'];
-        },
         isNavbar() {
             if (this.$route.path.includes('/wallet')) {
                 return false;
