+++ src/components/wallet/portfolio/FungibleRow.vue
