+++ src/router/index.ts
@@ -2,14 +2,9 @@ import Vue from 'vue'
 import VueRouter, {Route} from 'vue-router'
 import Home from '../views/Home.vue'
 
-import Ava from '../views/wallet/Ava.vue';
-
-import Transfer from '@/components/wallet/transfer/Transfer.vue';
-// import Assets from '../views/wallet/Assets.vue';
-import History from '../views/wallet/History.vue';
-// import Transfer from '../views/wallet/Transfer.vue';
+import Transfer from '@/views/wallet/Transfer.vue';
 import Advanced from '@/components/wallet/advanced/Advanced.vue';
-import ManageKeys from '@/components/wallet/keys/ManageKeys.vue';
+import ManageKeys from '@/views/wallet/ManageKeys.vue';
 import Menu from '../views/access/Menu.vue';
 import AccessString from '../views/access/AccessString.vue';
 import Keystore from '../views/access/Keystore.vue';
@@ -17,12 +12,12 @@ import Mnemonic from "@/views/access/Mnemonic.vue";
 import Access from '../views/access/Access.vue';
 import Create from "@/views/Create.vue";
 import Wallet from "@/views/Wallet.vue";
+import WalletHome from "@/views/wallet/Portfolio.vue";
 
 Vue.use(VueRouter);
 
 
 import store from '../store/index' // your vuex store
-import WalletHome from "@/views/wallet/Home.vue";
 
 
 const ifNotAuthenticated = (to: Route, from: Route, next: Function) => {
@@ -94,10 +89,6 @@ const routes = [
             {
                 path: 'keys',
                 component: ManageKeys
-            },
-            {
-                path: 'advanced',
-                component: Advanced
             }
         ],
         component: Wallet,
