--- src/views/LoadingApp.vue
@@ -1,15 +1,0 @@
-<template>
-    <div class="loading_app">
-            Loading AVAX Wallet...
-    </div>
-</template>
-<style scoped lang="scss">
-    .loading_app{
-        width: 100%;
-        height: 100%;
-        background-color: #fff;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-</style>
